Move the class to a separate file
