+++ src/System.Windows.Forms/src/System/Windows/Forms/UpDownBase.UpDownEdit.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -11,7 +11,7 @@ namespace System.Windows.Forms
 {
     public abstract partial class UpDownBase
     {
-        internal class UpDownEdit : TextBox
+        internal partial class UpDownEdit : TextBox
         {
             private readonly UpDownBase _parent;
             private bool _doubleClickFired;
@@ -108,24 +108,6 @@ namespace System.Windows.Forms
 
             protected override void OnLostFocus(EventArgs e)
                 => _parent.InvokeLostFocus(_parent, e);
-
-            internal class UpDownEditAccessibleObject : ControlAccessibleObject
-            {
-                readonly UpDownBase _parent;
-
-                public UpDownEditAccessibleObject(UpDownEdit owner, UpDownBase parent) : base(owner)
-                {
-                    _parent = parent;
-                }
-
-                public override string Name
-                {
-                    get => _parent.AccessibilityObject.Name;
-                    set => _parent.AccessibilityObject.Name = value;
-                }
-
-                public override string KeyboardShortcut => _parent.AccessibilityObject.KeyboardShortcut;
-            }
         }
     }
 }
